polyfill for requestIdleCallback that also detects react-native
