+++ src/bugcatch.js
@@ -1,26 +1,6 @@
 "use strict";
 import { initVitals } from "./vitals";
-import { xhrPost, newEvent } from "./api";
-
-window.requestIdleCallback =
-    window.requestIdleCallback ||
-    function (cb) {
-        var start = Date.now();
-        return setTimeout(function () {
-            cb({
-                didTimeout: false,
-                timeRemaining: function () {
-                    return Math.max(0, 50 - (Date.now() - start));
-                },
-            });
-        }, 1);
-    };
-
-window.cancelIdleCallback =
-    window.cancelIdleCallback ||
-    function (id) {
-        clearTimeout(id);
-    };
+import { post, newEvent } from "./api";
 
 /**
  * Default options object
